+++ sql/planbuilder/builder.go
@@ -170,6 +170,14 @@ func (s *scope) copy() *scope {
 	return &ret
 }
 
+func DeepCopyNode(node sql.Node) (sql.Node, error) {
+	n, _, err := transform.NodeExprs(node, func(e sql.Expression) (sql.Expression, transform.TreeIdentity, error) {
+		e, err := transform.Clone(e)
+		return e, transform.NewTree, err
+	})
+	return n, err
+}
+
 func (s *scope) addCte(name string, cteScope *scope) {
 	if s.ctes == nil {
 		s.ctes = make(map[string]*scope)
@@ -607,6 +615,18 @@ func (b *PlanBuilder) buildDataSource(inScope *scope, te ast.TableExpr) (outScop
 	return
 }
 
+func columnsToStrings(cols ast.Columns) []string {
+	if len(cols) == 0 {
+		return nil
+	}
+	res := make([]string, len(cols))
+	for i, c := range cols {
+		res[i] = c.String()
+	}
+
+	return res
+}
+
 func (b *PlanBuilder) buildAsOf(inScope *scope, asOf ast.Expr) interface{} {
 	var err error
 	asOfExpr := b.buildScalar(inScope, asOf)
@@ -617,7 +637,7 @@ func (b *PlanBuilder) buildAsOf(inScope *scope, asOf ast.Expr) interface{} {
 	return asOfLit
 }
 
-func (b *PlanBuilder) buildResolvedTable(tab, db string, asOf interface{}) *plan.ResolvedTable {
+func (b *PlanBuilder) resolveTable(tab, db string, asOf interface{}) *plan.ResolvedTable {
 	table, _, err := b.cat.TableAsOf(b.ctx, db, tab, asOf)
 	if err != nil {
 		b.handleErr(err)
