+++ sql/planbuilder/select.go
@@ -1,0 +1,30 @@
+package planbuilder
+
+import (
+	"github.com/dolthub/go-mysql-server/sql"
+	"github.com/dolthub/go-mysql-server/sql/expression"
+	"github.com/dolthub/go-mysql-server/sql/plan"
+	ast "github.com/dolthub/vitess/go/vt/sqlparser"
+)
+
+func selectExprNeedsAlias(e *ast.AliasedExpr, expr sql.Expression) bool {
+	if len(e.InputExpression) == 0 {
+		return false
+	}
+
+	// We want to avoid unnecessary wrapping of aliases, but not at the cost of blowing up parse time. So we examine
+	// the expression tree to see if is likely to need an alias without first serializing the expression being
+	// examined, which can be very expensive in memory.
+	complex := false
+	sql.Inspect(expr, func(expr sql.Expression) bool {
+		switch expr.(type) {
+		case *plan.Subquery, *expression.UnresolvedFunction, *expression.Case, *expression.InTuple, *plan.InSubquery, *expression.HashInTuple:
+			complex = true
+			return false
+		default:
+			return true
+		}
+	})
+
+	return complex || e.InputExpression != expr.String()
+}
